requirements.txt
aiogram==2.25.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
apscheduler==3.10.4